Moved installation tasks to installation service.
